arm: efi: add a dummy efi arch on ARM64
this arch will forces a generic arm 64 cpu, which will
produce an efi payload on ARM 64, this works with a virtual
machine using qemu-system-aarch64, need further work to support
real hardware.
Submit the patch in case someone is interrested in working
to have a full support for efi payload on real hardware.

Signed-off-by: Chali Anis <chalianis1@gmail.com>